fixed saving price

+++ src/BP.Api/BP.Domain/Entities/Product.cs
@@ -7,7 +7,7 @@ public record Product : ITableEntity
 {
     public string? ProductName { get; set; }
     public string SKU { get; set; } = null!;
-    public decimal Price { get; set; }
+    public double Price { get; set; }
     public int Stock { get; set; }
     public string PartitionKey { get; set; } = null!;
     public string RowKey { get; set; } = null!;
@@ -16,5 +16,5 @@ public record Product : ITableEntity
     public string MaterialCode { get; set; } = null!;
     public string CollectionCode { get; set; } = null!;
     public string FeatureCodes { get; set; } = null!;
-    public double YearCode { get; set; }
+    public int YearCode { get; set; }
 }
